% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootMargin.R
\name{bootMargin}
\alias{bootMargin}
\title{Significant Causes and Effects For Complete Graphs}
\usage{
bootMargin(CC, reps = 10000, pValue = 0.05, thr = 0.5, delete = FALSE)
}
\arguments{
\item{CC}{Three-dimensional matrix, where each submatrix along the z-axis is a square and reflective incidence matrix,
or a \code{list} of \code{data.frames} containing square and reflective incidence matrices.}

\item{reps}{Number of bootstrap replicas. By default \code{reps = 10,000}.}

\item{pValue}{Real between [0,1]. Significance threshold of the left one-sided mean t-test. By default \code{pValue = 0.05}.}

\item{thr}{\code{Real} between [0,1]. Defines the degree of significant truth. By default \code{thr = 0.5}.}

\item{delete}{\code{Logical}. For significant = TRUE, remove the rows and columns whose mean occurrences are not significant at the set p-value.
By default \code{significant = FALSE}.}
}
\value{
For \code{delete = TRUE}, it returns the three-dimensional matrix entered, but with the non-significant rows and columns
removed and a "directEffects" \code{data.frame} containing the following values:
\item{Var}{Variable name}
\item{Mean}{Calculated mean}
\item{ICU}{Upper Confidence Interval}
\item{p-value}{The p-value for the t test}
}
\description{
Performs the calculation of the mean incidence for each cause and each effect for all experts,
 the left one-sided confidence interval and the p-value. The function makes it possible to eliminate causes
 and effects whose average incidence is not significant at the established p-value.
}
\details{
The function implements "boot.one.bca" from the wBoot package to get the UCI and the p-value.
}
\examples{
# Example of use
# saveData <- bootMargin(CC = m$AA, reps = 1000, pValue = 0.05, thr = 0.5, delete = TRUE)
# To see the new data set:
##    saveData$Data
# To see the results obtained by row:
##    savaData$MarginRow
# To see the results obtained by col:
##    savaData$MarginCol
}
