% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCent_nuevo.R
\name{centrality.sq}
\alias{centrality.sq}
\title{Significant Centrality For Complete Graphs}
\usage{
centrality.sq(
  CE,
  Model = c("conpl", "conlnorm", "mean"),
  parallel = c("multicore", "snow", "no"),
  reps = 10000,
  ncpus = 1,
  conf = 0.95
)
}
\arguments{
\item{CE}{Three-dimensional matrix, where each submatrix along the z-axis is a square and reflective incidence matrix, or a \code{list} of \code{data.frames} containing square and reflective incidence matrices.}

\item{Model}{Bootstrap with one of the following statistics: \code{"conpl","conlnorm","mean"}. By default \code{Model = "mean"}.}

\item{parallel}{The type of parallel operation that is used (if applicable). The options are \code{"multicore", "snow" and "no"}. By default \code{parallel = "no"}.}

\item{reps}{Number of bootstrap replicas. By default \code{reps = 10,000}.}

\item{ncpus}{\code{Integer}. Number of processes that are used in the parallel implementation. By default \code{ncpus = 1}.}

\item{conf}{\code{Real}. Indicates the confidence levels of the required intervals. By default \code{conf = 0.95}.}
}
\value{
Returns a data.frame containing the following:
\item{Var}{  Variable name.}
\item{Median}{ Calculated median.}
\item{LCI}{  Lower confidence interval.}
\item{UCI}{  Upper confidence interval.}
\item{Method}{ Statistical method used.}
}
\description{
Perform the calculation of the median betweenness centrality with R bootstrap replicas for bipartite graphs.
}
\details{
The Model parameter makes use of the PoweRlaw package. For \code{“conpl”} the median of a power distribution is calculated according to Newman, M. E. (2005)
, or \code{"conlnorm"} can be used according to Gillespie CS (2015). In the event that either of the two statistical methods fails,
the error will be reported and the mean centrality will be calculated.

The parallel and ncpus options are not available on Windows operating systems.
}
\examples{
# a <- centrality.rect( CC = m$AA, CE = AB, EE = BB, Model = "conpl")
}
\references{
Canty A, Ripley BD (2021). boot: Bootstrap R (S-Plus) Functions. R package version 1.3-28.

Csardi G, Nepusz T (2006). "The igraph software package for complex network research." InterJournal, Complex Systems, 1695

Gillespie CS (2015). "Fitting Heavy Tailed Distributions: The poweRlaw Package." Journal of Statistical Software, 64(2), 1-16.

Newman, M. E. (2005). Power laws, Pareto distributions and Zipf's law. Contemporary physics, 46(5), 323-351.
}
