% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createdata.R
\name{createDataSet}
\alias{createDataSet}
\title{createDataSet}
\usage{
createDataSet(rows, columns, experts)
}
\arguments{
\item{rows}{Number of rows}

\item{columns}{Number of columns}

\item{experts}{Number of experts}
}
\value{
returns a list containing 3 three-dimensional matrices
}
\description{
createDataSet
}
