% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directEffects.r
\name{de.sq}
\alias{de.sq}
\title{Significant Direct Effects For Complete Graphs}
\usage{
de.sq(CC, reps = 10000, pValue = 0.05, thr = 0.5, delete = FALSE)
}
\arguments{
\item{CC}{Three-dimensional matrix, where each submatrix along the z-axis is a square and reflective incidence matrix,
or a \code{list} of \code{data.frames} containing square and reflective incidence matrices.}

\item{reps}{Number of bootstrap replicas. By default \code{reps = 10,000}.}

\item{pValue}{Real between [0,1]. Significance threshold of the left one-sided mean t-test. By default \code{pValue = 0.05}.}

\item{thr}{\code{Real} between [0,1]. Defines the degree of significant truth. By default \code{thr = 0.5}.}

\item{delete}{\code{Logical}. For significant = TRUE, remove the rows and columns whose mean occurrences are not significant at the set p-value.
By default \code{significant = FALSE.}}
}
\value{
For \code{delete = TRUE}, it returns the three-dimensional matrix entered, but with the non-significant rows and columns
removed and a "directEffects" \code{data.frame} containing the following values:
\item{From}{Origin of the incident}
\item{To}{Destination of the incident}
\item{Mean}{Average incidence}
\item{UCI}{Upper Confidence Interval}
\item{p-value}{The p-value for the t test}
For \code{delete = FALSE}, only the \code{data.frame} "directEffects" is returned.
}
\description{
Performs the calculation of the mean incidence, left one-sided confidence interval and p-value
 with multiple experts. The function allows to eliminate the edges whose mean incidences are not significant
 at the set p-value.
}
\details{
The function implements "boot.one.bca" from the wBoot package to get the UCI and the p-value.
}
\examples{
# saveData <- de.sq(CC =m$AA, reps = 1000, pValue = 0.05, thr = 0.5, delete = TRUE)
}
