% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DJ_d}
\alias{DJ_d}
\title{Dow Jones daily}
\format{A tibble with 18838 rows and 3 variables:
\describe{
  \item{Date}{dates in the format "mmddyy".}
  \item{Dow_Jones}{daily Dow Jones Industial Average prices.}
  \item{r_Dow_Jones}{log returns of the daily Dow Jones Industial Average prices.}
}}
\usage{
DJ_d
}
\description{
Daily Dow Jones Industrial Average.
18839 daily observations of closing prices, spanning 01/1915-02/1990.
}
\keyword{datasets}
