% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DJ_w}
\alias{DJ_w}
\title{Dow Jones weekly}
\format{A tibble with 4686 rows and 5 variables:
\describe{
  \item{Date}{dates in the format "mmddyy".}
  \item{High}{weekly Dow Jones Industial Average high.}
  \item{Low}{weekly Dow Jones Industial Average low.}
  \item{Close}{weekly Dow Jones Industial Average close prices.}
  \item{r_close}{log returns of the weekly Dow Jones Industial Average close prices.}
}}
\usage{
DJ_w
}
\description{
Weekly Dow Jones Industrial Average.
4686 weekly observations of high, low and closing prices, spanning 01/1900-10/1989.
}
\keyword{datasets}
