% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\name{EstCAPM}
\alias{EstCAPM}
\title{Estimate the CAPM model}
\usage{
EstCAPM(mZ, vZm)
}
\arguments{
\item{mZ}{matrix of the excess returns for N assets with dimension T by N.}

\item{vZm}{vector of the excess returns of the market portfolio.}
}
\value{
a list containing the following results:
\item{alpha}{estimated alpha.}
\item{beta}{estimated market beta.}
\item{mE}{matrix of the residuals.}
\item{Sigma}{estimated covariance matrix of the errors.}
}
\description{
This function estimates the Sharpe-Lintner CAPM model as follows:
\deqn{Z_t = \alpha + \beta Z_{mt} + \varepsilon_t.}
where \eqn{Z_t} is the excess returns for N assets,
\eqn{Z_{mt}} is the excess return of the market portfolio.
}
\examples{
mR = portfolio_m[,25:124]
vRm = portfolio_m[,3]
vRf = portfolio_m[,4]

mZ = mR - c(vRf)
vZm = vRm - c(vRf)

ret = EstCAPM(mZ, vZm)

}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{capm}
