% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factors_m}
\alias{factors_m}
\title{SMB and HML factor as constructed by Fama and French (1993).}
\format{A tibble with 942 rows and 6 variables:}
\usage{
factors_m
}
\description{
SMB is the difference between the returns on a small-stock portfolio (based on the low 30% quantile of market capitalization) and the returns on a big-stock portfolio (based on the top 30% quantile of market capitalization).
}
\details{
HML is the difference between the returns on a low-book-to-market portfolio (based on the low 30% quantile of the book-to-market equity ratio) and the returns on a high-book-to-market portfolio (based on the top 30% quantile of the book-to-market equity ratio).

Moreover, it includes the value-weighted return on all NYSE, AMEX, and NASDAQ stocks (from CRSP) as a proxy for the market as well as the one-month Treasury bill rate.

Note that all returns are measured in percent!

942 monthly observations, 07/1926-12/2004.

\itemize{
  \item year, year in integers.
  \item month, month in integers
  \item Market, value-weighted return on all NYSE, AMEX, and NASDAQ stocks (from CRSP) as a proxy for the market return.
  \item Tbill, one-month Treasury bill rate.
  \item SMB, difference between the returns on a small-stock portfolio (based on the low 30% quantile of market capitalization) and the returns on a big-stock portfolio (based on the top 30% quantile of market capitalization).
  \item HML, difference between the returns on a low-book-to-market portfolio (based on the low 30% quantile of the book-to-market equity ratio) and the returns on a high-book-to-market portfolio (based on the top 30% quantile of the book-to-market equity ratio).
}
}
\keyword{datasets}
