% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE.R
\docType{package}
\name{FE}
\alias{FE}
\alias{FE-package}
\title{FE: A package for the master and PhD course "Financial Econometrics".}
\description{
The package consisting of all the data sets and functions used in the master and PhD course "Financial Econometrics".
}
\section{Author and Maintainer}{

Yukai Yang

Department of Statistics, Uppsala University

\email{yukai.yang@statistik.uu.se}
}

