% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index_d}
\alias{index_d}
\title{Daily international stock indexes}
\format{A tibble with 3128 rows and 7 variables:
\describe{
  \item{DAXINDX}{German DAX}
  \item{FRCAC40}{Paris}
  \item{FTSE100}{London}
  \item{HNGKNGI}{Hong Kong}
  \item{NIKKEI}{Tokyo}
  \item{SNGALLS}{Singapore All Shares}
  \item{SPCOMP}{S&P500 New York}
}}
\usage{
index_d
}
\description{
3128 daily observations of the following stock indexes with some missing observations, spanning 06/01/86 to 31/12/97.
}
\keyword{datasets}
