% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Compute the log-likelihood function based on the residuals from a regression.}
\usage{
LogLikelihood(mE)
}
\arguments{
\item{mE}{a T by N matrix of residuals obtained from a regression.}
}
\value{
log-likelihood function.
}
\description{
This function computes the log-likelihood function based on a matrix of residuals obtained from a regression.
}
\details{
The assumptions of the regression model contain the multivariate normality of the error vectors.

It computes the following function value
\deqn{-\frac{NT}{2}\log 2 \pi - \frac{T}{2}\log |\hat{\Sigma}| - \frac{1}{2} \sum_{t=1}^T e_t' \hat{\Sigma} e_t}
where \eqn{e_t} is the residual vector at time t (the t's row of \code{mE}),
and \eqn{\hat{\Sigma}} is the covariance matrix of the residuals (the estimated covariance matrix of the errors).
}
\examples{
mR = portfolio_m[,25:124]
vRm = portfolio_m[,3]
vRf = portfolio_m[,4]

mZ = mR - c(vRf)
vZm = vRm - c(vRf)

## get the residuals
mE = EstCAPM(mZ, vZm)$mE

LogLikelihood(mE)

}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{log-likelihood}
