% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{PCA}
\alias{PCA}
\title{Principal component analysis}
\usage{
PCA(mx, num = 1)
}
\arguments{
\item{mx}{matrix of variables, rows are observations and columns are variables.}

\item{num}{number of factors that you need.}
}
\value{
a list of the results.
\item{factors}{matrix of the factors found in PCA, rows are observations, columns are factors.}
\item{weights}{weights that produce the factors from the variables.}
\item{explained}{how much do the factors explain the variation of the variables, in percentage.}
}
\description{
Principal component analysis
}
\examples{

}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{factor}
