% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{portcap_m}
\alias{portcap_m}
\title{Value-weighted returns for size portfolios}
\format{A tibble with 942 rows and 21 variables:}
\usage{
portcap_m
}
\description{
Value-weighted returns for size portfolios based on stocks traded at the NYSE and AMEX.
}
\details{
The portfolios are reconstructed every June based on market capitalization (cap).
The data set contains the returns of portfolios associated
with the low 30%, mid 40%, top 30% quantile as well as with 5 quintiles and 10 deciles.

Moreover, it includes the value-weighted return on all NYSE, AMEX, and NASDAQ stocks (from CRSP) as a proxy for the market return as well as the one-month Treasury bill rate.

Note that all returns are measured in percent!

942 monthly observations, spanning 07/1926-12/2004.

\itemize{
  \item Date, dates in the format "yyyymm".
  \item Lo_30, portfolio returns for low 30\%.
  \item Mid_40, portfolio returns for mid 40\%.
  \item Hi_30, portfolio returns for top 30\%.
  \item Lo_20, portfolio returns for low 20\%.
  \item Hi_20, portfolio returns for top 20\%.
  \item Lo_10, portfolio returns for low 10\%.
  \item Hi_10, portfolio returns for top 10\%.
  \item Xyy.yy, portfolio returns for yy-yy\%.
  \item Rf, risk-free one-month Treasury bill returns.
  \item Market, the value-weighted return on all NYSE, AMEX, and NASDAQ stocks (from CRSP) as a proxy for the market return.
}
}
\keyword{datasets}
