% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\name{SharpeLintner}
\alias{SharpeLintner}
\title{Implement the Sharpe-Lintner's version of the CAPM}
\usage{
SharpeLintner(mu, Omega, rf, mu_p = NULL)
}
\arguments{
\item{mu}{a vector of returns of some assets.}

\item{Omega}{a positive definite covariance matrix of the returns.}

\item{rf}{a risk-free return.}

\item{mu_p}{a scalar of the expected return of some portfolio, by default \code{NULL}.}
}
\value{
a list containing the following results:
\item{w_m}{the weights of the market portfolio.}
\item{mu_m}{the expected return of the market portfolio.}
\item{var_m}{the variance of the market portfolio.}
\item{w_p}{the weights of the portfolio which achieves the return \code{mu_p} together with a risk-free asset with return \code{rf}.}
\item{mu_p}{the expected return of the portfolio.}
\item{var_p}{the variance of the portfolio.}
}
\description{
This function implements the Sharpe-Lintner's version of the CAPM model.
}
\details{
The function takes the arguments of some returns and their covariance matrix, and a risk-free return.
It returns a list containing the weights, expected return and variance of the market portfolio or the tangency portfolio.

If some expected return \code{mu_p} is provided, then it also returns the weights and variance of the portfolio which achieves it.
}
\examples{
Omega = crossprod(matrix(rnorm(25),5,5))
mu = runif(5,min=.05,max=.15)

SharpeLintner(mu, Omega, .02)

}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{capm}
